pyproject.toml
[project]
name = "parts-multiagent"
version = "0.1.0"
description = "Symmetric LAN A2A agents that route natural language inventory questions across local CSV data folders."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pandas>=2.2.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["parts_multiagent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
